+++ CHANGELOG.md
@@ -3,6 +3,9 @@
 
 ## 2.4.0 (not yet released)
 
+### Added
+ - (experimental) A `create_api` function to generate an API similar to the one of `portion` but configured to use a given subclass of `Interval`.
+
 ### Changed
  - Speed up `repr` and `to_string` for `Interval` instances (see [#76](https://github.com/AlexandreDecan/portion/issues/76), adm271828).
  - Some internal changes to ease subclassing: